Extract set_active_pixels from  add_energy_channel
Every call to 'add_energy_channel_mask_allpanel_cuda' would convert the
pixel_mask to a list and copy the list to the GPU. Since the mask does
not change, this only needs to be done once before the first call. This
is now possible with the new 'set_active_pixels' function. As this adds
the list to the detector, it should probably be moved to the detector
class.